+++ TimberCommons/CommonUIPatches/ManufactoryDescriberGetCraftingTimePatch.cs
@@ -9,7 +9,7 @@ using Timberborn.Localization;
 using Timberborn.WorkshopsUI;
 
 // ReSharper disable InconsistentNaming
-namespace IgorZ.TimberCommons.CommonUI {
+namespace IgorZ.TimberCommons.CommonUIPatches {
 
 /// <summary>Harmony patch to show days and hours for the slow recipes.</summary>
 [HarmonyPatch(typeof(ManufactoryDescriber), nameof(ManufactoryDescriber.GetCraftingTime))]